pyproject.toml
[project]
name = "document-intelligence-refinery"
version = "0.1.0"
description = "Production-grade agentic pipeline for document extraction with provenance"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pdfplumber>=0.11.0",
    "docling>=2.0.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
